[refactor] remove Layout from index.tsx
