+++ renderer/components/containers/Layout.tsx
@@ -10,13 +10,11 @@ import ContextBridge from './ContextBridge'
 
 type Props = {
   children: ReactNode
-  title?: string
 }
 
-const Layout = ({ children, title = 'こんとれ！！' }: Props) => (
+const Layout = ({ children }: Props) => (
   <StyledConteiner>
     <Head>
-      <title>{title}</title>
       <meta charSet="utf-8" />
       <meta name="viewport" content="initial-scale=1.0, width=device-width" />
     </Head>
